Tear out all of the useless parametric types.
The idea with the parametric types was to avoid the allocation costs of having a
Union type in structs, but that cost is minimized in 0.7.0 (coming soon tm).
Additionally, if these structs were to be packed into an array, they'd have to
be an abstract type or union type anyway, so the gain was minimal in the first
place.