Remove complexity
It's not being replaced with anything yet, but anything that requires
that much cruft before I can even use it to login to matrix isn't a
good design. In the future, I may end up back at a similar place, but
I want to build up to that complexity rather than trying to front-load
it and ending up with vestigial weirdness.